+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Migrate to [magnus](https://github.com/matsadler/magnus)
+
 ## [0.1.0] - 2023-02-21
 
 - Initial release